refactor: Separate the trailing system library into 2D trailing, 3D trailing, and common trailing. 重构: 分离拖尾系统库到2D拖尾、3D拖尾，公用拖尾.
