+++ gulpfile.js
@@ -292,6 +292,24 @@ const packsDef = [{
         './layaAir/laya/TileMap_Discarded/**/**.ts'
     ],
 },
+{
+    'libName': "trailCommon",
+    'input': [
+        './layaAir/laya/trail/trailCommon/**/*.*',
+    ],
+},
+{
+    'libName': "trail2D",
+    'input': [
+        './layaAir/laya/trail/trail2D/**/*.*',
+    ],
+},
+{
+    'libName': "trail3D",
+    'input': [
+        './layaAir/laya/trail/trail3D/**/*.*',
+    ],
+},
 ];
 
 /*
