+++ src/layaAir/laya/trail/trail2D/ModuleDef.ts
@@ -1,0 +1,5 @@
+import { ClassUtils } from "../../utils/ClassUtils";
+import { Trail2DRender } from "./Trail2DRender";
+let c = ClassUtils.regClass;
+
+c("Trail2DRender", Trail2DRender);
