+++ src/layaAir/laya/trail/trail3D/ModuleDef.ts
@@ -1,0 +1,10 @@
+import { ClassUtils } from "../../utils/ClassUtils";
+import { TrailFilter } from "./TrailFilter";
+import { TrailMaterial } from "./TrailMaterial";
+import { TrailRenderer } from "./TrailRenderer";
+
+let c = ClassUtils.regClass;
+
+c("TrailFilter", TrailFilter);
+c("TrailRenderer", TrailRenderer);
+c("TrailMaterial", TrailMaterial);
