+++ src/layaAir/laya/trail/trailCommon/RenderFeatureComman/Trail/TrailShaderCommon.ts
@@ -1,8 +1,8 @@
-import { LayaGL } from "../../../layagl/LayaGL";
-import { Color } from "../../../maths/Color";
-import { Vector4 } from "../../../maths/Vector4";
-import { ShaderDataType } from "../../../RenderDriver/DriverDesign/RenderDevice/ShaderData";
-import { Shader3D } from "../../../RenderEngine/RenderShader/Shader3D";
+import { LayaGL } from "../../../../layagl/LayaGL";
+import { Color } from "../../../../maths/Color";
+import { Vector4 } from "../../../../maths/Vector4";
+import { ShaderDataType } from "../../../../RenderDriver/DriverDesign/RenderDevice/ShaderData";
+import { Shader3D } from "../../../../RenderEngine/RenderShader/Shader3D";
 import TrailVertexUtilGLSL from "./Shader/TrailVertexUtil.glsl";
 import { VertexTrail } from "./VertexTrail";
