+++ src/layaAir/laya/trail/trailCommon/RenderFeatureComman/Trail/TrailTextureMode.ts
